\name{xplain-package}
\alias{xplain-package}
\docType{package}

\title{
xplain Package
}

\description{
This package allows to provide live interpretations and explanations of statistical functions in R. These interpretations and explanations are shown when the explained function is called by the user. They can interact with the values of the explained function's actual results to offer relevant, meaningful insights. The xplain interpretations and explanations are based on an easy-to-use XML format that allows to include R code to interact with the returns of the explained function.
}

\author{
Author and maintainer: Joachim Zuckarelli, \email{joachim@zuckarelli.de}.
}

\section{Ressources on the internet}{
Web tutorial on how to work with xplain: \url{https://www.zuckarelli.de/xplain/index.html}\cr
xplain cheat sheet: \url{https://www.zuckarelli.de/xplain/xplain_cheatsheet.pdf}\cr
xplain on GitHub: \url{https://www.github.com/jsugarelli/xplain}
}

\section{Comments and suggestions}{
Your comments and suggestions are highly appreciated.
}

\seealso{
\code{\link{xplain}}, \code{\link{xplain.overview}}, \code{\link{xplain.getcall}}
}
